+++ src/main/kotlin/com/ccalecs/contacorrente/response/TransferenceResponse.kt
@@ -1,11 +1,12 @@
 package com.ccalecs.contacorrente.response
 
+import java.math.BigDecimal
 import java.time.LocalDate
 
 data class TransferenceResponse(
     val transferenceId: Long?,
     val clientIdSender: Long?,
     val clientIdReceiver: Long?,
-    val value: Double,
+    val value: BigDecimal,
     val data: LocalDate
 )
